+++ ft_strncmp.c
@@ -6,16 +6,18 @@
 /*   By: sadawi <sadawi@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/15 14:57:01 by sadawi            #+#    #+#             */
-/*   Updated: 2019/10/23 16:41:41 by sadawi           ###   ########.fr       */
+/*   Updated: 2019/10/29 13:33:56 by sadawi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strncmp(const char *s1, const char *s2, int n)
+#include "libft.h"
+
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
 	int i;
 
 	i = 0;
-	while ((s1[i] || s2[i]) && i < n)
+	while ((s1[i] || s2[i]) && i < (int)n)
 	{
 		if ((unsigned char)s1[i] != (unsigned char)s2[i])
 			return ((unsigned char)s1[i] - (unsigned char)s2[i]);
